+++ src/main/java/atbmhttt/atbmcq_16/client/Views/Subjects/SubjectsView.java
@@ -1,12 +1,11 @@
 package atbmhttt.atbmcq_16.client.Views.Subjects;
 
-import atbmhttt.atbmcq_16.MonHocController;
 import javafx.scene.Scene;
 import javafx.scene.layout.BorderPane;
 
 public class SubjectsView {
-    public void displaySubjects(BorderPane contentArea, String username, String password) {
-        MonHocController controller = new MonHocController(username, password);
+    public void displaySubjects(BorderPane contentArea) {
+        MonHocController controller = new MonHocController();
         Scene scene = controller.createScene();
         contentArea.setCenter(scene.getRoot());
     }
